Add check for INS Fix before building GPS/Odom msg.
Otherwise the CompositeData object throws an error with a VN100 in the callback which doesn't come to the surface while running the rosnode but does cause a slowdown so the max IMU rate is closer to 40Hz. The issue arises from calling cd.insStatus() before checking that the insStatus exists.